+++ index.php
@@ -23,7 +23,7 @@
           </div>
         </article>
       <?php endwhile; ?>
-        <?php zeroTheme_blogPagination(); ?>
+        <?php blogPagination(); ?>
       <?php else : ?>
         <h2>Not Found</h2>
       <?php endif; ?>