+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.2] - 2021-08-16
+### Removed
+* Removed local FontAwesome instance
+* Restored default WordPress comments
+
+### Added
+* New default logo, new `screenshot.png` file
+* Additional headers for `style.css`, including preview information for WordPress
+
+### Changed
+* Improved comments
+* Removed unnecessary comments, attributes and variables
+* Set `Theme URL` in `style.css` as default GitHub repo URL for `plugin-update-checker`
+
+### Updated
+* Updated `YahnisElsts/plugin-update-checker` to v4.11
+
 ## [0.1] - 2019-07-04
 ### Initial commit